@@ -1,30 +1,83 @@
 import { useState, useCallback } from 'react';
 
-export interface Toast {
-  id: string;
-  message: string;
-  type: 'info' | 'success' | 'warning' | 'error';
+type ToastType = 'success' | 'error' | 'warning' | 'info';
+type ToastPosition =
+  | 'top-right'
+  | 'top-left'
+  | 'bottom-right'
+  | 'bottom-left'
+  | 'top-center'
+  | 'bottom-center';
+
+interface ToastOptions {
+  type?: ToastType;
+  position?: ToastPosition;
   duration?: number;
 }
 
-export function useToast() {
+interface Toast {
+  id: string;
+  message: string;
+  type: ToastType;
+  position: ToastPosition;
+}
+
+interface UseToastReturn {
+  toasts: Toast[];
+  addToast: (message: string, options?: ToastOptions) => void;
+  removeToast: (id: string) => void;
+  clearToasts: () => void;
+}
+
+/**
+ * React hook for managing toast notifications
+ * @param defaultPosition - Default position for toasts
+ * @returns Toast management functions
+ */
+export function useToast(
+  defaultPosition: ToastPosition = 'top-right'
+): UseToastReturn {
   const [toasts, setToasts] = useState<Toast[]>([]);
 
+  const addToast = useCallback(
+    (message: string, options: ToastOptions = {}) => {
+      const id = Math.random().toString(36).substring(2, 9);
+      const {
+        type = 'info',
+        position = defaultPosition,
+        duration = 3000,
+      } = options;
+
+      const newToast: Toast = {
+        id,
+        message,
+        type,
+        position,
+      };
+
+      setToasts(prev => [...prev, newToast]);
+
+      if (duration > 0) {
+        setTimeout(() => {
+          removeToast(id);
+        }, duration);
+      }
+    },
+    [defaultPosition]
+  );
+
   const removeToast = useCallback((id: string) => {
-    setToasts((prev) => prev.filter((t) => t.id !== id));
+    setToasts(prev => prev.filter(toast => toast.id !== id));
   }, []);
 
-  const addToast = useCallback((message: string, type: Toast['type'] = 'info', duration = 3000) => {
-    const id = Math.random().toString(36).substr(2, 9);
-    const newToast = { id, message, type, duration };
-    setToasts((prev) => [...prev, newToast]);
+  const clearToasts = useCallback(() => {
+    setToasts([]);
+  }, []);
 
-    if (duration > 0) {
-      setTimeout(() => {
-        removeToast(id);
-      }, duration);
-    }
-  }, [removeToast]);
-
-  return { toasts, addToast, removeToast };
+  return {
+    toasts,
+    addToast,
+    removeToast,
+    clearToasts,
+  };
 }
