@@ -1,43 +1,46 @@
-import { useState } from 'react';
+import { useState, useCallback } from 'react';
 
-interface UseCounterOptions {
-  min?: number;
-  max?: number;
-}
-
-interface UseCounterReturn {
-  count: number;
+interface CounterActions {
   increment: () => void;
   decrement: () => void;
-  set: (value: number) => void;
   reset: () => void;
+  setValue: (value: number) => void;
 }
 
-export function useCounter(initialValue = 0, options: UseCounterOptions = {}): UseCounterReturn {
-  const { min, max } = options;
-  const [count, setCount] = useState(initialValue);
+/**
+ * Counter hook for numeric state management
+ * @param initialValue - Initial counter value (default: 0)
+ * @param step - Step increment/decrement value (default: 1)
+ * @returns [value, actions] tuple
+ */
+export function useCounter(
+  initialValue = 0,
+  step = 1
+): [number, CounterActions] {
+  const [count, setCount] = useState<number>(initialValue);
 
-  const increment = () => {
-    setCount(c => {
-      const next = c + 1;
-      return max !== undefined && next > max ? c : next;
-    });
-  };
+  const increment = useCallback(() => {
+    setCount(prev => prev + step);
+  }, [step]);
 
-  const decrement = () => {
-    setCount(c => {
-      const next = c - 1;
-      return min !== undefined && next < min ? c : next;
-    });
-  };
+  const decrement = useCallback(() => {
+    setCount(prev => prev - step);
+  }, [step]);
 
-  const set = (value: number) => {
-    setCount(value);
-  };
-
-  const reset = () => {
+  const reset = useCallback(() => {
     setCount(initialValue);
+  }, [initialValue]);
+
+  const setValue = useCallback((value: number) => {
+    setCount(value);
+  }, []);
+
+  const actions: CounterActions = {
+    increment,
+    decrement,
+    reset,
+    setValue,
   };
 
-  return { count, increment, decrement, set, reset };
+  return [count, actions];
 }
