@@ -1,39 +1,50 @@
-import { ref, readonly } from 'vue';
+import { ref, Ref } from 'vue';
 
-interface UseCounterOptions {
-  min?: number;
-  max?: number;
+interface CounterActions {
+  increment: () => void;
+  decrement: () => void;
+  reset: () => void;
+  setValue: (value: number) => void;
 }
 
-export function useCounter(initialValue = 0, options: UseCounterOptions = {}) {
-  const count = ref(initialValue);
-  const { min, max } = options;
+interface UseCounterReturn {
+  value: Ref<number>;
+  increment: () => void;
+  decrement: () => void;
+  reset: () => void;
+  setValue: (value: number) => void;
+}
+
+/**
+ * Counter hook for numeric state management
+ * @param initialValue - Initial counter value (default: 0)
+ * @param step - Step increment/decrement value (default: 1)
+ * @returns Object with value ref and action methods
+ */
+export function useCounter(initialValue = 0, step = 1): UseCounterReturn {
+  const value: Ref<number> = ref(initialValue);
 
   const increment = () => {
-    const next = count.value + 1;
-    if (max !== undefined && next > max) return;
-    count.value = next;
+    value.value += step;
   };
 
   const decrement = () => {
-    const next = count.value - 1;
-    if (min !== undefined && next < min) return;
-    count.value = next;
-  };
-
-  const set = (value: number) => {
-    count.value = value;
+    value.value -= step;
   };
 
   const reset = () => {
-    count.value = initialValue;
+    value.value = initialValue;
+  };
+
+  const setValue = (newValue: number) => {
+    value.value = newValue;
   };
 
   return {
-    count: readonly(count),
+    value,
     increment,
     decrement,
-    set,
-    reset
+    reset,
+    setValue,
   };
 }
