@@ -1,14 +1,35 @@
-import { Injectable, ElementRef } from '@angular/core';
-import { fromEvent, merge, Observable } from 'rxjs';
-import { map } from 'rxjs/operators';
+import { ElementRef, signal, WritableSignal } from '@angular/core';
 
-@Injectable({ providedIn: 'root' })
-export class HoverService {
-  // Usually this is done via (mouseenter) and (mouseleave) in template
-  // But to be consistent with "hook" style:
-  observe(element: ElementRef): Observable<boolean> {
-    const enter$ = fromEvent(element.nativeElement, 'mouseenter').pipe(map(() => true));
-    const leave$ = fromEvent(element.nativeElement, 'mouseleave').pipe(map(() => false));
-    return merge(enter$, leave$);
+/**
+ * Track hover state for an element.
+ * In Angular, use the useHoverRef directive or call useHover() with an ElementRef.
+ * @param elementRef - Optional ElementRef to attach to
+ * @returns Object with isHovered signal and methods to bind (or use directive)
+ */
+export function useHover<T extends HTMLElement = HTMLElement>(
+  elementRef?: ElementRef<T>
+): { isHovered: WritableSignal<boolean>; setElement: (el: T | null) => void } {
+  const isHovered = signal(false);
+  let currentEl: T | null = null;
+
+  const setElement = (el: T | null) => {
+    if (currentEl) {
+      currentEl.removeEventListener('mouseenter', onEnter);
+      currentEl.removeEventListener('mouseleave', onLeave);
+    }
+    currentEl = el;
+    if (el) {
+      el.addEventListener('mouseenter', onEnter);
+      el.addEventListener('mouseleave', onLeave);
+    }
+  };
+
+  const onEnter = () => isHovered.set(true);
+  const onLeave = () => isHovered.set(false);
+
+  if (elementRef?.nativeElement) {
+    setElement(elementRef.nativeElement);
   }
+
+  return { isHovered, setElement };
 }
