@@ -1,16 +1,21 @@
-import { Injectable } from '@angular/core';
+import { effect, signal, Signal } from '@angular/core';
 
-@Injectable({ providedIn: 'root' })
-export class PreviousService<T> {
-  private _previous: T | undefined;
-  private _current: T | undefined;
+/**
+ * Store and return the previous value of a signal
+ * @param value - Signal to track
+ * @returns Signal containing the previous value (undefined on first run)
+ */
+export function usePrevious<T>(value: Signal<T>): Signal<T | undefined> {
+  const previous = signal<T | undefined>(undefined);
+  let lastValue: T = value();
 
-  update(value: T) {
-    this._previous = this._current;
-    this._current = value;
-  }
+  effect(() => {
+    const current = value();
+    if (current !== lastValue) {
+      previous.set(lastValue);
+      lastValue = current;
+    }
+  });
 
-  get previous(): T | undefined {
-    return this._previous;
-  }
+  return previous;
 }
