@@ -1,15 +1,18 @@
 import { useState, useCallback } from 'react';
 
-export function useToggle(initialValue: boolean = false): [boolean, (nextValue?: any) => void] {
+/**
+ * Toggle hook for boolean state management
+ * @param initialValue - Initial boolean value (default: false)
+ * @returns [value, toggle, setValue] tuple
+ */
+export function useToggle(
+  initialValue = false
+): [boolean, () => void, (value: boolean) => void] {
   const [value, setValue] = useState<boolean>(initialValue);
 
-  const toggle = useCallback((nextValue?: any) => {
-    if (typeof nextValue === 'boolean') {
-      setValue(nextValue);
-    } else {
-      setValue(v => !v);
-    }
+  const toggle = useCallback(() => {
+    setValue(prev => !prev);
   }, []);
 
-  return [value, toggle];
+  return [value, toggle, setValue];
 }
